refactor: centralize post result construction
