+++ src/background/platform-poster.test.ts
@@ -4,7 +4,6 @@ import { threadsAdapter } from '../adapters/threads';
 import type { PlatformAdapter } from '../adapters/types';
 import {
   buildVerifyExpectationForChunk,
-  buildFinalChunkResult,
   buildDomPostAttempts,
   getComposeUrlForMedia,
   isAmbiguousPostDispatchError,
@@ -127,32 +126,6 @@ describe('platform poster helpers', () => {
     expect(shouldReuseExistingTabForAttempt(adapter(), false, {
       reuseExistingTab: false,
     })).toBe(false);
-  });
-
-  it('preserves the final chunk flow trace on aggregated preview results', () => {
-    const result = buildFinalChunkResult('x', false, true, undefined, {
-      mode: 'preview',
-      attempt: 'default',
-      submitReached: false,
-      lastCompletedStep: 'wait-submit',
-    });
-
-    expect(result.flow).toMatchObject({
-      mode: 'preview',
-      attempt: 'default',
-      submitReached: false,
-      lastCompletedStep: 'wait-submit',
-    });
-  });
-
-  it('adds a fallback flow trace when a successful chunk omitted it', () => {
-    const result = buildFinalChunkResult('x', false, true);
-
-    expect(result.flow).toMatchObject({
-      mode: 'preview',
-      submitReached: false,
-      lastCompletedStep: 'preview-flow',
-    });
   });
 
   it('expects media only on the first chunk of a split post', () => {
